Implemented MainActivity.kt PR's feedback
