+++ app/src/main/java/com/oracle/visualize/presentation/screens/mainScreen/MainView.kt
@@ -8,20 +8,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavDestination.Companion.hasRoute
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.toRoute
-import com.oracle.visualize.presentation.navigation.NavRoutes
 import com.oracle.visualize.presentation.components.BottomNavBar
-import com.oracle.visualize.presentation.screens.createChartScreen.CreatePage
-import com.oracle.visualize.presentation.screens.feedScreen.FeedPage
-import com.oracle.visualize.presentation.screens.notificationScreen.NotificationPage
-import com.oracle.visualize.presentation.screens.profileScreen.ProfilePage
-import com.oracle.visualize.presentation.screens.fullVisualizationScreen.FullVisualizationPage
-
+import com.oracle.visualize.presentation.navigation.AppNavHost
 
 /**
  * Main container screen that sets up the navigation host and bottom bar.
@@ -34,18 +24,20 @@ fun MainScreen(
 ) {
     val navController = rememberNavController()
     val backStackEntry by navController.currentBackStackEntryAsState()
+    val navItems = viewModel.navItems
 
-    val currentDestination = viewModel.navItems.find { item ->
+    val currentDestination = navItems.find { item ->
         backStackEntry?.destination?.hasRoute(item.destination::class) == true
     }?.destination
 
     val showBottomBar = currentDestination != null
 
     Scaffold(
+        modifier = Modifier.fillMaxSize(),
         bottomBar = {
             if (showBottomBar) {
                 BottomNavBar(
-                    navItems = viewModel.navItems,
+                    navItems = navItems,
                     currentDestination = currentDestination,
                     onItemSelected = { destination ->
                         navController.navigate(destination) {
@@ -64,59 +56,5 @@ fun MainScreen(
             navController = navController,
             modifier = Modifier.padding(innerPadding)
         )
-    }
-}
-
-// ─── NavHost ──────────────────────────────────────────────────────────────────
-
-@Composable
-fun AppNavHost(navController: NavHostController, modifier: Modifier = Modifier) {
-    NavHost(
-        navController = navController,
-        startDestination = NavRoutes.Feed,
-        modifier = modifier
-    ) {
-        composable<NavRoutes.Feed> {
-            FeedPage(
-                modifier = Modifier.fillMaxSize(),
-                onVisualizationClick = { visualizationId ->
-                    navController.navigate(
-                        NavRoutes.FullScreen(visualizationId)
-                    )
-                }
-            )
-        }
-
-        composable<NavRoutes.Create> {
-            CreatePage(modifier = Modifier.fillMaxSize())
-        }
-
-        composable<NavRoutes.Notifications> {
-            NotificationPage(modifier = Modifier.fillMaxSize())
-        }
-
-        composable<NavRoutes.Teams> {
-            // TODO: Implement TeamsPage
-        }
-
-        composable<NavRoutes.Profile> {
-            ProfilePage(modifier = Modifier.fillMaxSize())
-            // TODO: Pass profile.userId to ProfilePage
-        }
-
-        composable<NavRoutes.FullScreen> { backStackEntry ->
-            val route = backStackEntry.toRoute<NavRoutes.FullScreen>()
-
-            FullVisualizationPage(
-                visualizationId = route.visualizationId,
-                modifier = Modifier.fillMaxSize(),
-                onBackClick = {
-                    navController.popBackStack()
-                },
-                onThreadsClick = {
-                }
-            )
-        }
-
     }
 }
